backend/requirements.txt
blinker==1.9.0
click==8.2.2
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
redis==6.2.0
Werkzeug==3.1.3
